requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Werkzeug==2.3.7
WTForms==3.0.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Pillow==10.0.1
requests==2.31.0
# stripe==6.7.0  # Removido - solo MercadoPago
mercadopago==2.3.0
celery==5.3.1
redis==4.6.0
gunicorn==21.2.0



